Fixed the bug where the listview was not getting populated with the files already present in the recordings folder

+++ NameSayer/src/controllers/ListMenuController.java
@@ -32,9 +32,10 @@ public class ListMenuController implements Initializable {
     public ListView namesList;
     public ListView<NameVersions> namesVersion;
     public ListView selectedNames;
+
     private LinkedList<Names> nameObjects = new LinkedList<>();
-    private ObservableList<NameVersions> selectedVersionObjects = FXCollections.observableArrayList();
     final private ObservableList<String> namesViewList = FXCollections.observableArrayList();
+    private ObservableList<NameVersions> selectedVersionObjects = FXCollections.observableArrayList();
     final private ObservableList<String> selectedVersionsViewList = FXCollections.observableArrayList();
 
     Names tempName = null;
@@ -117,21 +118,42 @@ public class ListMenuController implements Initializable {
 
 
             File tempFolder = new File (currentWorkingDir + "/NameSayer/Recordings/" + tempName + "/");
+            File destination = new File(tempFolder + "/" + f.getName());
 
-            if (tempFolder.exists()) {
-                File destination = new File(tempFolder + "/" + f.getName());
-                Files.copy(f.toPath(), destination.toPath(), StandardCopyOption.REPLACE_EXISTING);
+            if (destination.exists()) {
+                boolean namePresent = false;
+                Names nameFound = null;
                 for (Names n: nameObjects) {
                     if (n.getName().equals(tempName)) {
-                        n.addVersion(tempName, destination.getAbsolutePath());
+//                        n.addVersion(tempName, destination.getAbsolutePath());
+                        namePresent = true;
+                        nameFound = n;
+                        break;
                     }
                 }
-            } else {
-                tempFolder.mkdirs();
-                File destination = new File(tempFolder + "/" + f.getName());
-                Files.copy(f.toPath(), destination.toPath(), StandardCopyOption.REPLACE_EXISTING);
-                nameObjects.add(new Names(tempName, destination.getAbsolutePath()));
 
+                if (namePresent) {
+                    nameFound.addVersion(tempName, destination.getAbsolutePath());
+                } else {
+                    nameObjects.add(new Names(tempName, destination.getAbsolutePath()));
+                }
+
+            } else {
+                if (tempFolder.exists()) {
+//                File destination = new File(tempFolder + "/" + f.getName());
+                    Files.copy(f.toPath(), destination.toPath(), StandardCopyOption.REPLACE_EXISTING);
+                    for (Names n: nameObjects) {
+                        if (n.getName().equals(tempName)) {
+                            n.addVersion(tempName, destination.getAbsolutePath());
+                        }
+                    }
+                } else {
+                    tempFolder.mkdirs();
+//                File destination = new File(tempFolder + "/" + f.getName());
+                    Files.copy(f.toPath(), destination.toPath(), StandardCopyOption.REPLACE_EXISTING);
+                    nameObjects.add(new Names(tempName, destination.getAbsolutePath()));
+
+                }
             }
 
         }
@@ -156,6 +178,7 @@ public class ListMenuController implements Initializable {
         }
 
         selectedVersionObjects.clear();
+        selectedVersionsViewList.clear();
         Main.loadMainPage();
     }
